Stop trying to support overload-server on Windows
