+++ quic/s2n-quic-bench/Cargo.toml
@@ -17,6 +17,7 @@ s2n-codec = { path = "../../common/s2n-codec", features = ["testing"] }
 s2n-quic-core = { path = "../s2n-quic-core", features = ["testing"] }
 
 # Dependencies for the overload-server binary
+[target.'cfg(unix)'.dependencies]
 clap = { version = "4", features = ["derive"] }
 mimalloc = "0.1"
 nix = { version = "0.31.1", features = ["socket"] }
